requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.1
cryptography==43.0.3
pynput==1.7.6
pygame==2.5.2
pillow==10.1.0
opencv-python==4.8.1.78
psutil==5.9.6
pyinstaller==6.3.0
